requirements.txt
boto3     >=1.9.197
pandas    >=0.25.0
psycopg2-binary  >=2.8.3
pyotp     >=2.3.0
pytz      >=2019.1
pyyaml    >=5.1
